fix(session): cancel autosave before surface session writes
